Rename navigation_constants.dart
