+++ lib/widgets/moviestar_nav_config.dart
@@ -29,7 +29,7 @@ import 'package:flutter/material.dart';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import 'package:moviestar/constants/navigation_constants.dart';
+import 'package:moviestar/constants/navigation.dart';
 import 'package:moviestar/providers/theme_provider.dart';
 import 'package:moviestar/widgets/solid_nav_models.dart';
 import 'package:moviestar/widgets/solid_nav_utils.dart';
