+++ lib/widgets/solid_nav_drawer.dart
@@ -27,7 +27,7 @@ library;
 
 import 'package:flutter/material.dart';
 
-import 'package:moviestar/constants/navigation_constants.dart';
+import 'package:moviestar/constants/navigation.dart';
 import 'package:moviestar/widgets/solid_nav_models.dart';
 
 /// A solid navigation drawer component.
