+++ lib/widgets/solid_navigation_manager.dart
@@ -27,7 +27,7 @@ library;
 
 import 'package:flutter/material.dart';
 
-import 'package:moviestar/constants/navigation_constants.dart';
+import 'package:moviestar/constants/navigation.dart';
 import 'package:moviestar/widgets/solid_nav_bar.dart';
 import 'package:moviestar/widgets/solid_nav_drawer.dart';
 import 'package:moviestar/widgets/solid_nav_models.dart';
